Fixing some issues with the FindBoost-alt.cmake file where the cmake code could not find boost if it was installed in a non-standard location. Correcting some OOP issues with some of the classes mainly by changing MXADataModel instances to IDataModel and IDataImport instances.

+++ src/Core/AttributeHelpers.h
@@ -74,7 +74,7 @@ class NumberAttributeHelper : public IAttributeHelper
     
     void convert(std::stringstream &sstream)
     {
-      int32 size = sizeof(T);
+      size_t size = sizeof(T);
       if (size != 1) {
         sstream << _value;
       } else {
